refactor: update numeric types from usize to i64 for consistency across API functions

+++ src/api/monsters.rs
@@ -14,8 +14,8 @@ use crate::{
 pub async fn get_all_monsters(
     settings: &Settings,
     drop: Option<ValidatedString>,
-    max_level: Option<usize>,
-    min_level: Option<usize>,
+    max_level: Option<i64>,
+    min_level: Option<i64>,
     name: Option<ValidatedStringWithSpaces>,
     pagination: Option<PaginationParams>,
 ) -> Result<serde_json::Value> {
